Circle refactor: sub-component for circle roles OK
